+++ index.ts
@@ -91,7 +91,7 @@ interface BlockType {
 
 // --------------------------------------------------------
 
-const VERSION = "0.9.3";
+const VERSION = "0.9.4";
 
 const tools: Tool[] = [
   {